+++ CHANGELOG.md
@@ -6,6 +6,8 @@ NEW FEATURES
    o Add --spikePercentFilter to fix the minimum percentage of reads which should align on spike-in genome (#23)
    o Add --keepSingleton option
    o Improve checkDesign function (#22)
+   o Add --noReadExtension parameters
+   o Add parameters for all mapper options
 
 BUG FIXES