+++ CHANGELOG.md
@@ -11,6 +11,7 @@ NEW FEATURES
 
 BUG FIXES
 
+   o Fix bug for Epic2 without inputs (#47)
    o Fix bugs in sample name (#39)
    o Fix bug in spike-in reads cleaning
    o Update BWA stats for PE data. Report pairs and not reads for mqc